+++ python/scripts/test.py
@@ -121,7 +121,7 @@ def test_index_udf(ndim: int, batch_size: int):
     signature = types.float32(
         types.CPointer(types.float32),
         types.CPointer(types.float32),
-        types.size_t, types.size_t)
+        types.uint64, types.uint64)
 
     @cfunc(signature)
     def python_dot(a, b, n, m):
